# storage_proxy: update view update backlog on correct shard when writing