--- a/db/legacy_schema_migrator.cc
+++ b/db/legacy_schema_migrator.cc
@@ -557,7 +557,7 @@ public:
                 db::schema_tables::add_type_to_schema_mutation(t.metadata, t.timestamp.time_since_epoch().count(), mutations);
             }
         }
-        return _qp.proxy().mutate_locally(std::move(mutations), tracing::trace_state_ptr());
+        return _qp.proxy().mutate_locally(std::move(mutations), tracing::trace_state_ptr()).discard_result();
     }
 
     future<> flush_schemas() {