--- a/db/view/node_view_update_backlog.hh
+++ b/db/view/node_view_update_backlog.hh
@@ -47,6 +47,7 @@ public:
     }
 
     update_backlog add_fetch(unsigned shard, update_backlog backlog);
+    update_backlog fetch_shard(unsigned shard);
 
     // Exposed for testing only.
     update_backlog load() const {