--- a/db/view/view_update_generator.hh
+++ b/db/view/view_update_generator.hh
@@ -110,7 +110,7 @@ public:
             flat_mutation_reader_v2&&,
             gc_clock::time_point);
 
-    future<std::optional<update_backlog>> generate_and_propagate_view_updates(const replica::table& table,
+    future<update_backlog> generate_and_propagate_view_updates(const replica::table& table,
             const schema_ptr& base,
             reader_permit permit,
             std::vector<view_and_base>&& views,