--- a/service/broadcast_tables/experimental/lang.cc
+++ b/service/broadcast_tables/experimental/lang.cc
@@ -116,7 +116,7 @@ future<query_result> execute_broadcast_table_query(
 
                 auto value = utf8_type->deserialize(q.new_value);
                 new_mutation.set_clustered_cell(clustering_key::make_empty(), "value", std::move(value), ts);
-                co_await proxy.mutate_locally(new_mutation, {}, {});
+                co_await proxy.mutate_locally(new_mutation, {}, {}).discard_result();
             }
 
             if (is_conditional) {