… service/paxos/paxos_state.cc
… service/raft/group0_state_machine.cc
… service/raft/raft_group0.cc
… service/storage_proxy.cc
… service/storage_proxy.hh
… service/storage_service.cc
… test/boost/batchlog_manager_test.cc
… test/boost/database_test.cc
… test/boost/repair_test.cc
… test/topology_custom/test_mv_backlog.py