@wmitros 2024-05-14
I added the missing perf_simple_query results so I'm changing this to a PR.