@wmitros 2024-05-15
To avoid an additional allocation, I moved the view update backlog getter (which updates the backlog) from an extra `then()` lower, as part of `database::apply`. As a result, `perf-simple-query` shows no degradation now (actually, I'm consistently getting ~200 insns/op fewer than without the patch)