@scylladb-promoter 2024-05-15
#### :red_circle: CI State: FAILURE

:x: - [Build](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/8817/)
#### Build Failure:
```
[2024-05-15T12:24:03.871Z] FAILED: build/dev/service/paxos/paxos_state.hh.o 
[2024-05-15T12:24:03.871Z] clang++ -MD -MT build/dev/service/paxos/paxos_state.hh.o -MF build/dev/service/paxos/paxos_state.hh.o.d -std=c++20 -I/jenkins/workspace/scylla-master/scylla-ci/scylla/seastar/include -I/jenkins/workspace/scylla-master/scylla-ci/scylla/build/dev/seastar/gen/include -U_FORTIFY_SOURCE -Werror=unused-result -fstack-clash-protection -DSEASTAR_API_LEVEL=7 -DSEASTAR_BUILD_SHARED_LIBS -DSEASTAR_SSTRING -DSEASTAR_ENABLE_ALLOC_FAILURE_INJECTION -DSEASTAR_LOGGER_COMPILE_TIME_FMT -DSEASTAR_SCHEDULING_GROUPS_COUNT=16 -DSEASTAR_LOGGER_TYPE_STDOUT -DSEASTAR_TYPE_ERASE_MORE -DBOOST_NO_CXX98_FUNCTION_BASE -DFMT_SHARED -I/usr/include/p11-kit-1 -std=gnu++20  -ffile-prefix-map=/jenkins/workspace/scylla-master/scylla-ci/scylla=. -march=westmere -DBOOST_ALL_DYN_LINK    -fvisibility=hidden -isystem abseil -Wall -Werror -Wextra -Wimplicit-fallthrough -Wno-mismatched-tags -Wno-c++11-narrowing -Wno-overloaded-virtual -Wno-unused-parameter -Wno-unsupported-friend -Wno-missing-field-initializers -Wno-deprecated-copy -Wno-psabi -Wno-enum-constexpr-conversion -Wno-error=deprecated-declarations -DXXH_PRIVATE_API -DSEASTAR_TESTING_MAIN -ffile-prefix-map=/jenkins/workspace/scylla-master/scylla-ci/scylla=. -march=westmere -DDEVEL -DSEASTAR_ENABLE_ALLOC_FAILURE_INJECTION -DSCYLLA_ENABLE_ERROR_INJECTION -DSCYLLA_ENABLE_PREEMPTION_SOURCE -O2 -DSCYLLA_BUILD_MODE=dev -iquote. -iquote build/dev/gen  --include service/paxos/paxos_state.hh -c -o build/dev/service/paxos/paxos_state.hh.o build/dev/gen/empty.cc
[2024-05-15T12:24:03.873Z] In file included from <built-in>:1:
[2024-05-15T12:24:03.873Z] ./service/paxos/paxos_state.hh:122:43: error: no member named update_backlog in namespace db::view
[2024-05-15T12:24:03.873Z]     static future<std::optional<db::view::update_backlog>> learn(storage_proxy& sp, db::system_keyspace& sys_ks, schema_ptr schema, proposal decision, clock_type::time_point timeout, tracing::trace_state_ptr tr_state);
[2024-05-15T12:24:03.874Z]                                 ~~~~~~~~~~^
[2024-05-15T12:24:03.874Z] 1 error generated.
```
#### Build Details:
- Duration: 2 hr 27 min
- Builder: i-06c1e5c5a955cecc7 (m5ad.8xlarge)