@scylladb-promoter 2024-05-15
#### :green_circle: CI State: SUCCESS

:white_check_mark: - [Build](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/8837/)
:white_check_mark: - [Unit Tests Custom](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/8837/)
**The following new/updated tests ran 100 times for each mode:**
:small_blue_diamond: boost/batchlog_manager_test
:small_blue_diamond: boost/database_test
:small_blue_diamond: boost/repair_test
:small_blue_diamond: topology_custom/test_mv_backlog
:white_check_mark: - [Container Test](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/8837/)
:white_check_mark: - [dtest with topology changes](https://jenkins.scylladb.com/job/scylla-master/job/gating-dtest-release-with-consistent-topology-changes/1592/)
:white_check_mark: - [dtest](https://jenkins.scylladb.com/job/scylla-master/job/gating-dtest-release/9614/)
:white_check_mark: - [Unit Tests](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/8837/)
#### Build Details:
- Duration: 8 hr 39 min
- Builder: i-0fba5819eb748e143 (m5ad.12xlarge)